==> 2 lsoa_data.ipynb <==
#Importing libraries
import pandas as pd

#Reading in CSV with mapping from LSOAs to postcodes
dfmap = pd.read_csv('input_data/pcd_lsoa_map.csv', dtype={'pcd': str, 'lsoa11' : str, 'lat' : float, 'long' : float})

#Selecting relevant columns
dfmap = dfmap[["pcd", "lsoa11", "lat", "long"]]

#Extracting short postcode
dfmap["pcd_adjusted"] = dfmap["pcd"].str[0:4]
dfmap["pcd_adjusted"] = dfmap["pcd_adjusted"].str.strip()

#Reading in CSVs with LSOA information
dflsoa1 = pd.read_csv('input_data/lsoa_data_1.csv', header = 0)
dflsoa2 = pd.read_csv('input_data/lsoa_data_2.csv', header = 0)

#Merging CSVs with LSOA information
dflsoa = pd.merge_ordered(dflsoa1, dflsoa2, how = "outer", on = "lsoa11")

#Transforming ages into percentages of total population
for key in dflsoa.columns:
    if "ages_" in key:
        dflsoa[key] = dflsoa[key] / dflsoa["all_ages"]
dflsoa = dflsoa.drop(labels = "all_ages", axis = 1).reset_index(drop=True)

#Transforming household types, ownership types, resident types, and dwelling types into percentages
for key in dflsoa.columns[7:22]:
    dflsoa[key] = dflsoa[key] / 100

#Transforming economic activity into percentages
total_economic = dflsoa['economic_active'] + dflsoa['economic_inactive']
for key in dflsoa.columns[23:28]:
    dflsoa[key] = dflsoa[key] / total_economic

#Transforming educational background, and health into percentages
for key in dflsoa.columns[28:38]:
    dflsoa[key] = dflsoa[key] / 100

#Merging LSOA information with postcode mapping
dflsoapcd = pd.merge_ordered(dfmap, dflsoa, how = "outer", on = "lsoa11")

#Aggregating information per postcode
dflsoapcdagg = dflsoapcd.groupby("pcd_adjusted").mean()
dflsoapcdagg = dflsoapcdagg.reset_index()

#Saving to CSV
dflsoapcdagg.to_csv("output_data/pcd_info.csv")
