==> 3 merged_data.ipynb <==
#Importing libraries
import pandas as pd
import numpy as np

#Reading in CSVs
rightmove = pd.read_csv("output_data/rentals_rightmove.csv", index_col = 0)
zoopla = pd.read_csv("output_data/rentals_zoopla.csv", index_col = 0)
pcd = pd.read_csv("output_data/pcd_info.csv", index_col = 0)

#Saving source of data
rightmove["source"] = "rightmove"
zoopla["source"] = "zoopla"

#Merging listings data
rentals = pd.concat([rightmove, zoopla], axis = 0, ignore_index = True)

#Mapping listings data to postcode information
df = pd.merge_ordered(rentals, pcd, how = "left", left_on = "postcode", right_on = "pcd_adjusted")

#Dropping listings where postcode cannot be mapped
df.dropna(inplace = True)

#Cleaning data
df["type"] = df["type"].apply(lambda x: x.lower())
type_dict = {'end of terrace': 'end terrace house', 
             'terraced': 'terraced house',
             'semi-detached': 'semi-detached house', 
             'mews': 'mews house', 
             'detached': 'detached house', 
             'ground flat': 'flat', 
             'equestrian property': 'property', 
             'town house': 'mews house'}
df["type"] = df["type"].replace(type_dict)

df["postcode"] = df["postcode"].apply(lambda x: x.strip())

#Dropping one postcode column
df.drop("pcd_adjusted", axis = 1, inplace = True)
df = df.reset_index(drop = True)

#Dropping economic activity, given more granular split in economically inactive, employee, self-employed, unemployed, student
df.drop("economic_active", axis = 1, inplace = True)
df = df.reset_index(drop = True)

#Saving to CSV 
df.to_csv("output_data/data_vf.csv")
